requirements.txt
PyYAML==6.0
Telethon==1.24.0
pysocks
